+++ src/constants/colors.js
@@ -1,0 +1,26 @@
+export default {
+  black: '#000',
+  white: '#fff',
+  gray: {
+    gray15: '#262626',
+    gray25: '#404040',
+    gray40: '#666666',
+    gray50: '#808080',
+    gray64: '#A4A4A4',
+    gray60: 'rgba(0, 0, 0, 0.6)',
+    gray70: '#B3B3B3',
+    gray96: '#F6F6F6',
+    gray97: '#F7F7F7',
+    gray98: '#F9F9F9',
+    lightGray: '#F1F1F1'
+  },
+  red: {
+    lightRed: '#D92616',
+    rustyRed: '#c71b0a',
+    liverRed: '#C7000A'
+  },
+  photographyColor: '#08192d',
+  primaryColor: '#c71b0a',
+  secondaryColor: '#a67a44'
+}
+